requirements.txt
Flask==2.3.3
psutil==5.9.5
py-cpuinfo==9.0.0
